Move metrics conversion to pkg/translate
