+++ exporter/prometheusremotewriteexporter/exporter.go
@@ -29,6 +29,7 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/prometheusremotewrite"
 	"github.com/prometheus/prometheus/prompb"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/confighttp"
@@ -105,7 +106,7 @@ func (prwe *prwExporter) PushMetrics(ctx context.Context, md pdata.Metrics) erro
 	case <-prwe.closeChan:
 		return errors.New("shutdown has been called")
 	default:
-		tsMap, dropped, errs := prwe.ConvertMetrics(ctx, md)
+		tsMap, dropped, errs := prometheusremotewrite.MetricsToPRW(prwe.namespace, prwe.externalLabels, md)
 
 		if exportErrors := prwe.export(ctx, tsMap); len(exportErrors) != 0 {
 			dropped = md.MetricCount()
@@ -118,74 +119,6 @@ func (prwe *prwExporter) PushMetrics(ctx context.Context, md pdata.Metrics) erro
 
 		return nil
 	}
-}
-
-func (prwe *PRWExporter) ConvertMetrics(ctx context.Context, md pdata.Metrics) (tsMap map[string]*prompb.TimeSeries, dropped int, errs error) {
-	tsMap = make(map[string]*prompb.TimeSeries)
-
-	resourceMetricsSlice := md.ResourceMetrics()
-	for i := 0; i < resourceMetricsSlice.Len(); i++ {
-		resourceMetrics := resourceMetricsSlice.At(i)
-		resource := resourceMetrics.Resource()
-		instrumentationLibraryMetricsSlice := resourceMetrics.InstrumentationLibraryMetrics()
-		// TODO: add resource attributes as labels, probably in next PR
-		for j := 0; j < instrumentationLibraryMetricsSlice.Len(); j++ {
-			instrumentationLibraryMetrics := instrumentationLibraryMetricsSlice.At(j)
-			metricSlice := instrumentationLibraryMetrics.Metrics()
-
-			// TODO: decide if instrumentation library information should be exported as labels
-			for k := 0; k < metricSlice.Len(); k++ {
-				metric := metricSlice.At(k)
-
-				// check for valid type and temporality combination and for matching data field and type
-				if ok := validateMetrics(metric); !ok {
-					dropped++
-					errs = multierr.Append(errs, consumererror.NewPermanent(errors.New("invalid temporality and type combination")))
-					continue
-				}
-
-				// handle individual metric based on type
-				switch metric.DataType() {
-				case pdata.MetricDataTypeGauge:
-					dataPoints := metric.Gauge().DataPoints()
-					if err := prwe.addNumberDataPointSlice(dataPoints, tsMap, resource, metric); err != nil {
-						dropped++
-						errs = multierr.Append(errs, err)
-					}
-				case pdata.MetricDataTypeSum:
-					dataPoints := metric.Sum().DataPoints()
-					if err := prwe.addNumberDataPointSlice(dataPoints, tsMap, resource, metric); err != nil {
-						dropped++
-						errs = multierr.Append(errs, err)
-					}
-
-				case pdata.MetricDataTypeHistogram:
-					dataPoints := metric.Histogram().DataPoints()
-					if dataPoints.Len() == 0 {
-						dropped++
-						errs = multierr.Append(errs, consumererror.NewPermanent(fmt.Errorf("empty data points. %s is dropped", metric.Name())))
-					}
-					for x := 0; x < dataPoints.Len(); x++ {
-						addSingleHistogramDataPoint(dataPoints.At(x), resource, metric, prwe.namespace, tsMap, prwe.externalLabels)
-					}
-				case pdata.MetricDataTypeSummary:
-					dataPoints := metric.Summary().DataPoints()
-					if dataPoints.Len() == 0 {
-						dropped++
-						errs = multierr.Append(errs, consumererror.NewPermanent(fmt.Errorf("empty data points. %s is dropped", metric.Name())))
-					}
-					for x := 0; x < dataPoints.Len(); x++ {
-						addSingleSummaryDataPoint(dataPoints.At(x), resource, metric, prwe.namespace, tsMap, prwe.externalLabels)
-					}
-				default:
-					dropped++
-					errs = multierr.Append(errs, consumererror.NewPermanent(errors.New("unsupported metric type")))
-				}
-			}
-		}
-	}
-
-	return
 }
 
 func validateAndSanitizeExternalLabels(externalLabels map[string]string) (map[string]string, error) {
@@ -205,16 +138,6 @@ func validateAndSanitizeExternalLabels(externalLabels map[string]string) (map[st
 	}
 
 	return sanitizedLabels, nil
-}
-
-func (prwe *prwExporter) addNumberDataPointSlice(dataPoints pdata.NumberDataPointSlice, tsMap map[string]*prompb.TimeSeries, resource pdata.Resource, metric pdata.Metric) error {
-	if dataPoints.Len() == 0 {
-		return consumererror.NewPermanent(fmt.Errorf("empty data points. %s is dropped", metric.Name()))
-	}
-	for x := 0; x < dataPoints.Len(); x++ {
-		addSingleNumberDataPoint(dataPoints.At(x), resource, metric, prwe.namespace, tsMap, prwe.externalLabels)
-	}
-	return nil
 }
 
 // export sends a Snappy-compressed WriteRequest containing TimeSeries to a remote write endpoint in order
