+++ pkg/translator/prometheusremotewrite/go.mod
@@ -1,0 +1,20 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/prometheusremotewrite
+
+go 1.17
+
+require (
+	github.com/prometheus/common v0.32.1
+	github.com/prometheus/prometheus v1.8.2-0.20220117154355-4855a0c067e2
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/collector v0.42.0
+	go.opentelemetry.io/collector/model v0.42.0
+	go.uber.org/multierr v1.7.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gogo/protobuf v1.3.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+)
